+++ pkg/api/query/v1.go
@@ -43,9 +43,7 @@ import (
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/util/annotations"
 	"github.com/prometheus/prometheus/util/stats"
-	promqlapi "github.com/thanos-io/promql-engine/api"
 	"github.com/thanos-io/promql-engine/engine"
-	"github.com/thanos-io/promql-engine/logicalplan"
 
 	"github.com/thanos-io/thanos/pkg/api"
 	"github.com/thanos-io/thanos/pkg/exemplars"
@@ -69,6 +67,7 @@ import (
 )
 
 const (
+	QueryParam               = "query"
 	DedupParam               = "dedup"
 	PartialResponseParam     = "partial_response"
 	MaxSourceResolutionParam = "max_source_resolution"
@@ -93,73 +92,13 @@ const (
 	PromqlEngineThanos     PromqlEngineType = "thanos"
 )
 
-type Engine interface {
-	MakeInstantQuery(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, qs string, ts time.Time) (promql.Query, error)
-	MakeRangeQuery(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, qs string, start, end time.Time, step time.Duration) (promql.Query, error)
-	MakeInstantQueryFromPlan(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, plan logicalplan.Node, ts time.Time) (promql.Query, error)
-	MakeRangeQueryFromPlan(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, root logicalplan.Node, start, end time.Time, step time.Duration) (promql.Query, error)
-}
-
-type prometheusEngineAdapter struct {
-	engine promql.QueryEngine
-}
-
-func (a *prometheusEngineAdapter) MakeInstantQuery(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, qs string, ts time.Time) (promql.Query, error) {
-	return a.engine.NewInstantQuery(ctx, q, opts, qs, ts)
-}
-
-func (a *prometheusEngineAdapter) MakeRangeQuery(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, qs string, start, end time.Time, step time.Duration) (promql.Query, error) {
-	return a.engine.NewRangeQuery(ctx, q, opts, qs, start, end, step)
-}
-
-func (a *prometheusEngineAdapter) MakeInstantQueryFromPlan(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, plan logicalplan.Node, ts time.Time) (promql.Query, error) {
-	return a.engine.NewInstantQuery(ctx, q, opts, plan.String(), ts)
-}
-
-func (a *prometheusEngineAdapter) MakeRangeQueryFromPlan(ctx context.Context, q storage.Queryable, opts *engine.QueryOpts, plan logicalplan.Node, start, end time.Time, step time.Duration) (promql.Query, error) {
-	return a.engine.NewRangeQuery(ctx, q, opts, plan.String(), start, end, step)
-}
-
-type QueryEngineFactory struct {
-	prometheus Engine
-	thanos     Engine
-}
-
-func (f *QueryEngineFactory) GetPrometheusEngine() Engine {
-	return f.prometheus
-}
-
-func (f *QueryEngineFactory) GetThanosEngine() Engine {
-	return f.thanos
-}
-
-func NewQueryEngineFactory(
-	engineOpts engine.Opts,
-	remoteEngineEndpoints promqlapi.RemoteEndpoints,
-) *QueryEngineFactory {
-	promEngine := promql.NewEngine(engineOpts.EngineOpts)
-	// set engine here to avoid duplicate registration of metrics
-	engineOpts.Engine = promEngine
-	var thanosEngine Engine
-	if remoteEngineEndpoints == nil {
-		thanosEngine = engine.New(engineOpts)
-	} else {
-		thanosEngine = engine.NewDistributedEngine(engineOpts, remoteEngineEndpoints)
-	}
-	return &QueryEngineFactory{
-		prometheus: &prometheusEngineAdapter{promEngine},
-		thanos:     thanosEngine,
-	}
-}
-
 // QueryAPI is an API used by Thanos Querier.
 type QueryAPI struct {
-	baseAPI         *api.BaseAPI
-	logger          log.Logger
-	gate            gate.Gate
-	queryableCreate query.QueryableCreator
-	// queryEngine returns appropriate promql.Engine for a query with a given step.
-	engineFactory       *QueryEngineFactory
+	baseAPI             *api.BaseAPI
+	logger              log.Logger
+	gate                gate.Gate
+	queryableCreate     query.QueryableCreator
+	queryFactory        *QueryFactory
 	defaultEngine       PromqlEngineType
 	lookbackDeltaCreate func(int64) time.Duration
 	ruleGroups          rules.UnaryClient
@@ -197,7 +136,7 @@ type QueryAPI struct {
 func NewQueryAPI(
 	logger log.Logger,
 	endpointStatus func() []query.EndpointStatus,
-	engineFactory *QueryEngineFactory,
+	queryFactory *QueryFactory,
 	defaultEngine PromqlEngineType,
 	lookbackDeltaCreate func(int64) time.Duration,
 	c query.QueryableCreator,
@@ -232,7 +171,7 @@ func NewQueryAPI(
 	return &QueryAPI{
 		baseAPI:                                api.NewBaseAPI(logger, disableCORS, flagsMap),
 		logger:                                 logger,
-		engineFactory:                          engineFactory,
+		queryFactory:                           queryFactory,
 		defaultEngine:                          defaultEngine,
 		lookbackDeltaCreate:                    lookbackDeltaCreate,
 		queryableCreate:                        c,
@@ -339,24 +278,22 @@ func (qapi *QueryAPI) parseEnableDedupParam(r *http.Request) (enableDeduplicatio
 	return enableDeduplication, nil
 }
 
-func (qapi *QueryAPI) parseEngineParam(r *http.Request) (queryEngine Engine, e PromqlEngineType, _ *api.ApiError) {
-	var engine Engine
+func (qapi *QueryAPI) parseQueryParam(r *http.Request) string {
+	return r.FormValue(QueryParam)
+}
 
-	param := PromqlEngineType(r.FormValue("engine"))
+func (qapi *QueryAPI) parseEngineParam(r *http.Request) (e PromqlEngineType, _ *api.ApiError) {
+	param := PromqlEngineType(r.FormValue(EngineParam))
 	if param == "" {
 		param = qapi.defaultEngine
 	}
-
 	switch param {
-	case PromqlEnginePrometheus:
-		engine = qapi.engineFactory.GetPrometheusEngine()
-	case PromqlEngineThanos:
-		engine = qapi.engineFactory.GetThanosEngine()
+	case PromqlEnginePrometheus, PromqlEngineThanos:
 	default:
-		return nil, param, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Errorf("'%s' bad engine", param)}
+		return param, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Errorf("'%s' bad engine", param)}
 	}
 
-	return engine, param, nil
+	return param, nil
 }
 
 func (qapi *QueryAPI) parseReplicaLabelsParam(r *http.Request) (replicaLabels []string, _ *api.ApiError) {
@@ -500,7 +437,7 @@ func processAnalysis(a *engine.AnalyzeOutputNode) queryTelemetry {
 }
 
 func (qapi *QueryAPI) queryExplain(r *http.Request) (interface{}, []error, *api.ApiError, func()) {
-	eng, engineParam, apiErr := qapi.parseEngineParam(r)
+	engineParam, apiErr := qapi.parseEngineParam(r)
 	if apiErr != nil {
 		return nil, nil, apiErr, func() {}
 	}
@@ -508,6 +445,7 @@ func (qapi *QueryAPI) queryExplain(r *http.Request) (interface{}, []error, *api.
 	if engineParam != PromqlEngineThanos {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: errors.New("engine type must be 'thanos'")}, func() {}
 	}
+	queryParam := qapi.parseQueryParam(r)
 
 	ts, err := parseTimeParam(r, "time", qapi.baseAPI.Now())
 	if err != nil {
@@ -565,18 +503,17 @@ func (qapi *QueryAPI) queryExplain(r *http.Request) (interface{}, []error, *api.
 	if lookbackDeltaFromReq > 0 {
 		lookbackDelta = lookbackDeltaFromReq
 	}
-
-	tenant, err := tenancy.GetTenantFromHTTP(r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField)
+	queryStr, _, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, queryParam)
 	if err != nil {
-		apiErr = &api.ApiError{Typ: api.ErrorBadData, Err: err}
-		return nil, nil, apiErr, func() {}
+		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
-	ctx = context.WithValue(ctx, tenancy.TenantKey, tenant)
 
-	var seriesStats []storepb.SeriesStatsCounter
-	qry, err := eng.MakeInstantQuery(
-		ctx,
-		qapi.queryableCreate(
+	var (
+		qry         promql.Query
+		seriesStats []storepb.SeriesStatsCounter
+	)
+	if err := tracing.DoInSpanWithErr(ctx, "instant_query_create", func(ctx context.Context) error {
+		queryable := qapi.queryableCreate(
 			enableDedup,
 			replicaLabels,
 			storeDebugMatchers,
@@ -585,16 +522,17 @@ func (qapi *QueryAPI) queryExplain(r *http.Request) (interface{}, []error, *api.
 			false,
 			shardInfo,
 			query.NewAggregateStatsReporter(&seriesStats),
-		),
-		&engine.QueryOpts{
+		)
+		queryOpts := &engine.QueryOpts{
 			LookbackDeltaParam:     lookbackDelta,
 			EnablePartialResponses: enablePartialResponse,
-		},
-		r.FormValue("query"),
-		ts,
-	)
+		}
 
-	if err != nil {
+		var qErr error
+		qry, qErr = qapi.queryFactory.makeInstantQuery(ctx, engineParam, queryable, planOrQuery{query: queryStr}, queryOpts, ts)
+		return qErr
+
+	}); err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
 
@@ -654,10 +592,11 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 		return nil, nil, apiErr, func() {}
 	}
 
-	eng, _, apiErr := qapi.parseEngineParam(r)
+	engineParam, apiErr := qapi.parseEngineParam(r)
 	if apiErr != nil {
 		return nil, nil, apiErr, func() {}
 	}
+	queryParam := qapi.parseQueryParam(r)
 
 	lookbackDelta := qapi.lookbackDeltaCreate(maxSourceResolution)
 	// Get custom lookback delta from request.
@@ -668,8 +607,7 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 	if lookbackDeltaFromReq > 0 {
 		lookbackDelta = lookbackDeltaFromReq
 	}
-
-	queryStr, tenant, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, r.FormValue("query"))
+	queryStr, tenant, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, queryParam)
 	if err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
@@ -678,28 +616,27 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 		qry         promql.Query
 		seriesStats []storepb.SeriesStatsCounter
 	)
+
 	if err := tracing.DoInSpanWithErr(ctx, "instant_query_create", func(ctx context.Context) error {
-		var err error
-		qry, err = eng.MakeInstantQuery(
-			ctx,
-			qapi.queryableCreate(
-				enableDedup,
-				replicaLabels,
-				storeDebugMatchers,
-				maxSourceResolution,
-				enablePartialResponse,
-				false,
-				shardInfo,
-				query.NewAggregateStatsReporter(&seriesStats),
-			),
-			&engine.QueryOpts{
-				LookbackDeltaParam:     lookbackDelta,
-				EnablePartialResponses: enablePartialResponse,
-			},
-			queryStr,
-			ts,
+		queryable := qapi.queryableCreate(
+			enableDedup,
+			replicaLabels,
+			storeDebugMatchers,
+			maxSourceResolution,
+			enablePartialResponse,
+			false,
+			shardInfo,
+			query.NewAggregateStatsReporter(&seriesStats),
 		)
-		return err
+		queryOpts := &engine.QueryOpts{
+			LookbackDeltaParam:     lookbackDelta,
+			EnablePartialResponses: enablePartialResponse,
+		}
+
+		var qErr error
+		qry, qErr = qapi.queryFactory.makeInstantQuery(ctx, engineParam, queryable, planOrQuery{query: queryStr}, queryOpts, ts)
+		return qErr
+
 	}); err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
@@ -751,7 +688,7 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 }
 
 func (qapi *QueryAPI) queryRangeExplain(r *http.Request) (interface{}, []error, *api.ApiError, func()) {
-	eng, engineParam, apiErr := qapi.parseEngineParam(r)
+	engineParam, apiErr := qapi.parseEngineParam(r)
 	if apiErr != nil {
 		return nil, nil, apiErr, func() {}
 	}
@@ -759,6 +696,7 @@ func (qapi *QueryAPI) queryRangeExplain(r *http.Request) (interface{}, []error, 
 	if engineParam != PromqlEngineThanos {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: errors.New("engine type must be 'thanos'")}, func() {}
 	}
+	queryParam := qapi.parseQueryParam(r)
 
 	start, err := parseTime(r.FormValue("start"))
 	if err != nil {
@@ -842,18 +780,17 @@ func (qapi *QueryAPI) queryRangeExplain(r *http.Request) (interface{}, []error, 
 	if lookbackDeltaFromReq > 0 {
 		lookbackDelta = lookbackDeltaFromReq
 	}
-
-	tenant, err := tenancy.GetTenantFromHTTP(r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField)
+	queryStr, _, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, queryParam)
 	if err != nil {
-		apiErr = &api.ApiError{Typ: api.ErrorBadData, Err: err}
-		return nil, nil, apiErr, func() {}
+		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
-	ctx = context.WithValue(ctx, tenancy.TenantKey, tenant)
 
-	var seriesStats []storepb.SeriesStatsCounter
-	qry, err := eng.MakeRangeQuery(
-		ctx,
-		qapi.queryableCreate(
+	var (
+		qry         promql.Query
+		seriesStats []storepb.SeriesStatsCounter
+	)
+	if err := tracing.DoInSpanWithErr(ctx, "range_query_create", func(ctx context.Context) error {
+		queryable := qapi.queryableCreate(
 			enableDedup,
 			replicaLabels,
 			storeDebugMatchers,
@@ -862,17 +799,17 @@ func (qapi *QueryAPI) queryRangeExplain(r *http.Request) (interface{}, []error, 
 			false,
 			shardInfo,
 			query.NewAggregateStatsReporter(&seriesStats),
-		),
-		&engine.QueryOpts{
+		)
+		queryOpts := &engine.QueryOpts{
 			LookbackDeltaParam:     lookbackDelta,
 			EnablePartialResponses: enablePartialResponse,
-		},
-		r.FormValue("query"),
-		start,
-		end,
-		step,
-	)
-	if err != nil {
+		}
+
+		var qErr error
+		qry, qErr = qapi.queryFactory.makeRangeQuery(ctx, engineParam, queryable, planOrQuery{query: queryStr}, queryOpts, start, end, step)
+		return qErr
+
+	}); err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
 
@@ -958,10 +895,11 @@ func (qapi *QueryAPI) queryRange(r *http.Request) (interface{}, []error, *api.Ap
 		return nil, nil, apiErr, func() {}
 	}
 
-	eng, _, apiErr := qapi.parseEngineParam(r)
+	engineParam, apiErr := qapi.parseEngineParam(r)
 	if apiErr != nil {
 		return nil, nil, apiErr, func() {}
 	}
+	queryParam := qapi.parseQueryParam(r)
 
 	lookbackDelta := qapi.lookbackDeltaCreate(maxSourceResolution)
 	// Get custom lookback delta from request.
@@ -972,46 +910,39 @@ func (qapi *QueryAPI) queryRange(r *http.Request) (interface{}, []error, *api.Ap
 	if lookbackDeltaFromReq > 0 {
 		lookbackDelta = lookbackDeltaFromReq
 	}
-
-	queryStr, tenant, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, r.FormValue("query"))
+	queryStr, tenant, ctx, err := tenancy.RewritePromQL(ctx, r, qapi.tenantHeader, qapi.defaultTenant, qapi.tenantCertField, qapi.enforceTenancy, qapi.tenantLabel, queryParam)
 	if err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
-
-	// Record the query range requested.
-	qapi.queryRangeHist.Observe(end.Sub(start).Seconds())
 
 	var (
 		qry         promql.Query
 		seriesStats []storepb.SeriesStatsCounter
 	)
 	if err := tracing.DoInSpanWithErr(ctx, "range_query_create", func(ctx context.Context) error {
-		var err error
-		qry, err = eng.MakeRangeQuery(
-			ctx,
-			qapi.queryableCreate(
-				enableDedup,
-				replicaLabels,
-				storeDebugMatchers,
-				maxSourceResolution,
-				enablePartialResponse,
-				false,
-				shardInfo,
-				query.NewAggregateStatsReporter(&seriesStats),
-			),
-			&engine.QueryOpts{
-				LookbackDeltaParam:     lookbackDelta,
-				EnablePartialResponses: enablePartialResponse,
-			},
-			queryStr,
-			start,
-			end,
-			step,
+		queryable := qapi.queryableCreate(
+			enableDedup,
+			replicaLabels,
+			storeDebugMatchers,
+			maxSourceResolution,
+			enablePartialResponse,
+			false,
+			shardInfo,
+			query.NewAggregateStatsReporter(&seriesStats),
 		)
-		return err
+		queryOpts := &engine.QueryOpts{
+			LookbackDeltaParam:     lookbackDelta,
+			EnablePartialResponses: enablePartialResponse,
+		}
+
+		var qErr error
+		qry, qErr = qapi.queryFactory.makeRangeQuery(ctx, engineParam, queryable, planOrQuery{query: queryStr}, queryOpts, start, end, step)
+		return qErr
+
 	}); err != nil {
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
+
 	analysis, err := qapi.parseQueryAnalyzeParam(r, qry)
 	if err != nil {
 		return nil, nil, apiErr, func() {}